requirements.txt
albumentations
numpy
opencv-python
pandas
Pillow
plotly
plyfile
psutil
pycocotools
pytorch
scipy
torch
torchvision
trimesh
ultralytics
pyyaml
